feat: group assignment for trusted header authentication
These changes make it possible to specify the groups an user
should be assigned to when trusted header authentication is
used.

The list of groups are read from a X-Forwarded-Remote-User-Groups
header. The name of the header follows similar naming convention
as the name header. The default header can be changed with the
REDASH_REMOTE_USER_GROUPS_HEADER setting.

The header should contain a comma separated list of group names
for the user. User will be assigned to each existing group
specified in the header. Non-existent groups found from the
header are ignored.

The feature is disabled by default and can be enabled with the
REDASH_REMOTE_USER_GROUPS_ENABLED setting.

Tests have also been written that exercise this functionality.